# v1/src/main/resources/static/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Escuela de Patinaje - API Completa
  version: "1.0.0"
  description: >-
    Documentación OpenAPI que cubre endpoints del frontend (vistas Thymeleaf)
    y del backend (APIs REST JSON) basados en el README del proyecto.
  contact:
    name: Equipo Escuela
    email: soporte@example.com
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /api/instructores:
    get:
      tags: [Instructores]
      summary: Listar instructores (API REST)
      responses:
        '200':
          description: Lista de instructores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Instructor'
    post:
      tags: [Instructores]
      summary: Crear instructor (API REST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorCreate'
      responses:
        '201':
          description: Instructor creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
  /api/instructores/{id}:
    get:
      tags: [Instructores]
      summary: Obtener instructor por id (API REST)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Instructor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
        '404':
          description: Instructor no encontrado
    put:
      tags: [Instructores]
      summary: Actualizar instructor (API REST)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Instructor'
      responses:
        '200':
          description: Instructor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instructor'
    delete:
      tags: [Instructores]
      summary: Eliminar instructor (API REST)
      responses:
        '200':
          description: Eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
...
  /api/asistencias:
    get:
      tags: [Asistencias]
      summary: Listar asistencias (API REST)
      responses:
        '200':
          description: Lista de asistencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asistencia'
    post:
      tags: [Asistencias]
      summary: Registrar asistencia (API REST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsistenciaCreate'
      responses:
        '201':
          description: Asistencia registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asistencia'
  /api/asistencias/alumno/{alumnoId}:
    get:
      tags: [Asistencias]
      summary: Obtener asistencias por alumno (API REST)
      parameters:
        - name: alumnoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de asistencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asistencia'
  /api/asistencias/clase/{claseId}:
    get:
      tags: [Asistencias]
      summary: Obtener asistencias por clase (API REST)
      parameters:
        - name: claseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de asistencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asistencia'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        fechaNacimiento:
          type: string
          format: date
        genero:
          type: string
        rol:
          type: string
          enum: [ADMIN, INSTRUCTOR, ALUMNO]
        activo:
          type: boolean
      example:
        id: 1
        name: "María Pérez"
        email: "maria@example.com"
        phone: "+34123456789"
        fechaNacimiento: "2005-09-15"
        genero: "Femenino"
        rol: "ALUMNO"
        activo: true
...
    Clase:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        nivel:
          type: string
        dia:
          type: string
        horaInicio:
          type: string
        horaFin:
          type: string
        capacidad:
          type: integer
        inscritos:
          type: integer
        descripcion:
          type: string
        estado:
          type: string
        instructor:
          $ref: '#/components/schemas/Instructor'
      example:
        id: 1
        nombre: "Patinaje Artístico"
        nivel: "Adultos"
        dia: "Lunes"
        horaInicio: "18:00"
        horaFin: "19:00"
        capacidad: 20
        inscritos: 5
        descripcion: "Clase para técnicas de patinaje artístico"
        estado: "Activa"
...
    Instructor:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        email:
          type: string
        telefono:
          type: string
        especialidad:
          type: string
        experiencia:
          type: string
        estado:
          type: string
        fechaRegistro:
          type: string
      example:
        id: 1
        nombre: "Carlos Gómez"
        email: "instructor@example.com"
        telefono: "+34123456789"
        especialidad: "Artístico"
        experiencia: "5 años"
        estado: "Activo"
        fechaRegistro: "2024-01-05"
    InstructorCreate:
      type: object
      required: [nombre, email]
      properties:
        nombre:
          type: string
        email:
          type: string
        telefono:
          type: string
        especialidad:
          type: string
        experiencia:
          type: string
    Asistencia:
      type: object
      properties:
        id:
          type: integer
        fecha:
          type: string
          format: date
        estado:
          type: string
        alumno:
          $ref: '#/components/schemas/User'
        clase:
          $ref: '#/components/schemas/Clase'
      example:
        id: 1
        fecha: "2025-12-05"
        estado: "Presente"
        alumno:
          id: 2
          name: "Ana"
        clase:
          id: 1
    AsistenciaCreate:
      type: object
      required: [fecha, estado, alumnoId, claseId]
      properties:
        fecha:
          type: string
          format: date
        estado:
          type: string
        alumnoId:
          type: integer
        claseId:
          type: integer